Fix compile errors in r24826.
TBR=bmeurer@chromium.org

Review URL: https://codereview.chromium.org/656793005

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24828 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/serialize.h
@@ -146,7 +146,7 @@ class AddressMapBase {
   }
 
   static uint32_t GetValue(HashMap::Entry* entry) {
-    return reinterpret_cast<uint32_t>(entry->value);
+    return static_cast<uint32_t>(reinterpret_cast<intptr_t>(entry->value));
   }
 
   static HashMap::Entry* LookupEntry(HashMap* map, HeapObject* obj,
